requirements/stable.txt
#Core
Django

# WSGI HTTP Server
gunicorn


# Debugging
ipdb
ipython
django-debug-toolbar


# Configuration
django-environ
pytz


# Models
psycopg2
Pillow

# Cache
django-redis


# Staticfiles
django-storages
boto


#Documentation
#mkdocs
djangorestframework

#testing
coverage

#logging



#format code
pep8
pyflakes
pylint






